cspad constants and data reading moved to serialtbx
